got compiling with nCtx configurable
